requirements.txt
torch
torchvision
numpy
Pillow
matplotlib
pandas
pyarrow
gradio 